cloud_function/kafka_notification/requirements.txt
functions-framework==3.*
snowflake-connector-python
KafkaProducer